+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+ - Generate and publish multilingual documentation from sfdx-hardis monitoring
+
 ## [5.36.3] 2025-05-21
 
 - Azure CI/CD workflows: use ubuntu-latest as default image