+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 
 - Generate and publish multilingual documentation from sfdx-hardis monitoring
 - Update command to install mkdocs-material & dependencies to match more python installation types
+- Upgrade way to call wrangler to publish to Cloudflare
 
 ## [5.36.3] 2025-05-21